Updated loading_screen and location_screen
Added loading icon, and passed data from loading_screen to location_screen by passing the data as dymanic; accessed the data on the loading screen using widget object which can then be used and manipulated.  Also created a String to change units of measurement to add at end of url; see OpenWeather API docs on how to change this.